+++ spine-ts/spine-phaser-v4/src/mixins.ts
@@ -24,7 +24,11 @@ SOFTWARE.
 
 // Adapted from https://github.com/agogpixel/phaser3-ts-utils/tree/main
 
-let components = (Phaser.GameObjects.Components as any);
+// biome-ignore-all lint: ignore biome for this file
+
+import * as Phaser from "phaser";
+
+const components = (Phaser.GameObjects.Components as any);
 export const ComputedSize = components.ComputedSize;
 export const Depth = components.Depth;
 export const Flip = components.Flip;