+++ spine-unity/Assets/Spine/Editor/spine-unity/Editor/Components/SkeletonGraphicInspector.cs
@@ -58,7 +58,7 @@ namespace Spine.Unity.Editor {
 		static GUIStyle ReloadButtonStyle { get { return EditorStyles.miniButton; } }
 
 		protected SerializedProperty material, color;
-		protected SerializedProperty additiveMaterial, multiplyMaterial, screenMaterial;
+		protected SerializedProperty additiveMaterial, multiplyMaterial, screenMaterial, forceAdditiveMaterial;
 		protected SerializedProperty skeletonDataAsset, initialSkinName;
 		protected SerializedProperty startingAnimation, startingLoop, timeScale, freeze,
 			updateTiming, updateWhenInvisible, unscaledTime, layoutScaleMode, editReferenceRect;
@@ -166,6 +166,7 @@ namespace Spine.Unity.Editor {
 			additiveMaterial = so.FindProperty("additiveMaterial");
 			multiplyMaterial = so.FindProperty("multiplyMaterial");
 			screenMaterial = so.FindProperty("screenMaterial");
+			forceAdditiveMaterial = so.FindProperty("forceAdditiveMaterial");
 
 			skeletonDataAsset = so.FindProperty("skeletonDataAsset");
 			initialSkinName = so.FindProperty("initialSkinName");
@@ -355,16 +356,22 @@ namespace Spine.Unity.Editor {
 
 									bool usesAdditiveMaterial = blendModeMaterials.applyAdditiveMaterial;
 									bool pmaVertexColors = thisSkeletonGraphic.MeshGenerator.settings.pmaVertexColors;
-									if (pmaVertexColors)
-										using (new EditorGUI.DisabledGroupScope(true)) {
-											EditorGUILayout.LabelField("Additive Material - Unused with PMA Vertex Colors", EditorStyles.label);
-										}
-									else if (usesAdditiveMaterial)
+									bool forceAdditiveEnabled = thisSkeletonGraphic.forceAdditiveMaterial;
+									if (pmaVertexColors) {
+										EditorGUILayout.PropertyField(forceAdditiveMaterial, SpineInspectorUtility.TempContent("Force Additive Material", null, "Still use 'Additive' material regardless of enabled 'PMA Vertex Colors'."));
+										if (forceAdditiveEnabled)
+											EditorGUILayout.PropertyField(additiveMaterial, SpineInspectorUtility.TempContent("Additive Material", null, "SkeletonGraphic Material for 'Additive' blend mode slots. Unused when 'PMA Vertex Colors' is enabled."));
+										else
+											using (new EditorGUI.DisabledGroupScope(true)) {
+												EditorGUILayout.LabelField("Additive Material - Unused with PMA Vertex Colors", EditorStyles.label);
+											}
+									} else if (usesAdditiveMaterial) {
 										EditorGUILayout.PropertyField(additiveMaterial, SpineInspectorUtility.TempContent("Additive Material", null, "SkeletonGraphic Material for 'Additive' blend mode slots. Unused when 'PMA Vertex Colors' is enabled."));
-									else
+									} else {
 										using (new EditorGUI.DisabledGroupScope(true)) {
 											EditorGUILayout.LabelField("No Additive Mat - 'Apply Additive Material' disabled at SkeletonDataAsset", EditorStyles.label);
 										}
+									}
 									EditorGUILayout.PropertyField(multiplyMaterial, SpineInspectorUtility.TempContent("Multiply Material", null, "SkeletonGraphic Material for 'Multiply' blend mode slots."));
 									EditorGUILayout.PropertyField(screenMaterial, SpineInspectorUtility.TempContent("Screen Material", null, "SkeletonGraphic Material for 'Screen' blend mode slots."));
 								}