+++ spine-unity/Modules/com.esotericsoftware.spine.ui-toolkit/Runtime/SpineVisualElement.cs
@@ -240,7 +240,7 @@ namespace Spine.Unity {
 			skeleton.UpdateWorldTransform(Skeleton.Physics.Update);
 		}
 
-		void Initialize (bool overwrite) {
+		public void Initialize (bool overwrite) {
 			if (this.IsValid && !overwrite) return;
 
 			if (this.SkeletonDataAsset == null) return;